requirements.txt
click==8.1.7
colorama==0.4.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
ruff==0.8.0
mysql-connector-python~=9.1.0